test: pin generated observe schema refs
Regenerate the checked-in tool definitions so artifact consumers receive the same self-contained observe schema as live tools/list.
